Remove "WithLimit" and "WithColumns", use default arguments instead

+++ src/main/scala/com/cognite/sdk/scala/common/Resource.scala
@@ -4,7 +4,7 @@ import com.cognite.sdk.scala.v1.RequestSession
 import com.softwaremill.sttp.Uri
 
 object Resource {
-  val defaultLimit: Long = 1000
+  val defaultLimit: Int = 1000
 }
 
 trait BaseUri {
